+++ tests/integration/CustomerTest.php
@@ -105,6 +105,103 @@ class CustomerTest extends Setup
         $this->assertNotNull($customer->merchantId);
     }
 
+    public function testCreateWithAccountTypeDebit()
+    {
+        $result = Braintree\Customer::create([
+            'firstName' => 'Mike',
+            'lastName' => 'Jones',
+            'company' => 'Jones Co.',
+            'email' => 'mike.jones@example.com',
+            'phone' => '419.555.1234',
+            'fax' => '419.555.1235',
+            'website' => 'http://example.com',
+            'creditCard' => [
+                'number' => Braintree\Test\CreditCardNumbers::$hiper,
+                'expirationDate' => '05/12',
+                'options' => [
+                    'verifyCard' => true,
+                    'verificationMerchantAccountId' => 'hiper_brl',
+                    'verificationAccountType' => 'debit'
+                ]
+            ]
+        ]);
+        $this->assertEquals(true, $result->success);
+        $this->assertEquals('debit', $result->customer->creditCards[0]->verification->creditCard['accountType']);
+    }
+
+    public function testCreateWithAccountTypeCredit()
+    {
+        $result = Braintree\Customer::create([
+            'firstName' => 'Mike',
+            'lastName' => 'Jones',
+            'company' => 'Jones Co.',
+            'email' => 'mike.jones@example.com',
+            'phone' => '419.555.1234',
+            'fax' => '419.555.1235',
+            'website' => 'http://example.com',
+            'creditCard' => [
+                'number' => Braintree\Test\CreditCardNumbers::$hiper,
+                'expirationDate' => '05/12',
+                'options' => [
+                    'verifyCard' => true,
+                    'verificationMerchantAccountId' => 'hiper_brl',
+                    'verificationAccountType' => 'credit'
+                ]
+            ]
+        ]);
+        $this->assertEquals(true, $result->success);
+        $this->assertEquals('credit', $result->customer->creditCards[0]->verification->creditCard['accountType']);
+    }
+
+    public function testCreateErrorsWithVerificationAccountTypeIsInvalid()
+    {
+        $result = Braintree\Customer::create([
+            'firstName' => 'Mike',
+            'lastName' => 'Jones',
+            'company' => 'Jones Co.',
+            'email' => 'mike.jones@example.com',
+            'phone' => '419.555.1234',
+            'fax' => '419.555.1235',
+            'website' => 'http://example.com',
+            'creditCard' => [
+                'number' => Braintree\Test\CreditCardNumbers::$hiper,
+                'expirationDate' => '05/12',
+                'options' => [
+                    'verifyCard' => true,
+                    'verificationMerchantAccountId' => 'hiper_brl',
+                    'verificationAccountType' => 'wrong'
+                ]
+            ]
+        ]);
+        $this->assertFalse($result->success);
+        $errors = $result->errors->forKey('customer')->forKey('creditCard')->forKey('options')->onAttribute('verificationAccountType');
+        $this->assertEquals(Braintree\Error\Codes::CREDIT_CARD_OPTIONS_VERIFICATION_ACCOUNT_TYPE_IS_INVALID, $errors[0]->code);
+    }
+
+    public function testCreateErrorsWithVerificationAccountTypeNotSupported()
+    {
+        $result = Braintree\Customer::create([
+            'firstName' => 'Mike',
+            'lastName' => 'Jones',
+            'company' => 'Jones Co.',
+            'email' => 'mike.jones@example.com',
+            'phone' => '419.555.1234',
+            'fax' => '419.555.1235',
+            'website' => 'http://example.com',
+            'creditCard' => [
+                "number" => "4111111111111111",
+                'expirationDate' => '05/12',
+                'options' => [
+                    'verifyCard' => true,
+                    'verificationAccountType' => 'credit'
+                ]
+            ]
+        ]);
+        $this->assertFalse($result->success);
+        $errors = $result->errors->forKey('customer')->forKey('creditCard')->forKey('options')->onAttribute('verificationAccountType');
+        $this->assertEquals(Braintree\Error\Codes::CREDIT_CARD_OPTIONS_VERIFICATION_ACCOUNT_TYPE_NOT_SUPPORTED, $errors[0]->code);
+    }
+
     public function testCreateCustomerWithCardUsingNonce()
     {
         $http = new HttpClientApi(Braintree\Configuration::$global);