+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## unreleased
 * Deprecate `GRANTED_PAYMENT_INSTRUMENT_UPDATE` and add `GRANTOR_UPDATED_GRANTED_PAYMENT_METHOD` and `RECIPIENT_UPDATED_GRANTED_PAYMENT_METHOD`
+* Add `accountType` field to `TransactionGateway`, `CreditCardGateway`, `PaymentMethodGateway`, and `CreditCardVerification`.
+* Add Hiper and Hipercard test numbers.
 
 ## 3.39.0
 * Add `bin` field to `paymentMethodNonceDetails`